perf: use concrete type for Counter, Gauge, and Histogram metrics

+++ collect/collect.go
@@ -268,7 +268,7 @@ func (i *InMemCollector) checkAlloc(ctx context.Context) {
 
 	var mem runtime.MemStats
 	runtime.ReadMemStats(&mem)
-	i.Metrics.Gauge("memory_heap_allocation", int64(mem.Alloc))
+	i.Metrics.Gauge("memory_heap_allocation", float64(mem.Alloc))
 	if maxAlloc == 0 || mem.Alloc < uint64(maxAlloc) {
 		return
 	}
@@ -300,7 +300,7 @@ func (i *InMemCollector) checkAlloc(ctx context.Context) {
 	// or just run out of traces to delete.
 
 	cacheSize := len(allTraces)
-	i.Metrics.Gauge("collector_cache_size", cacheSize)
+	i.Metrics.Gauge("collector_cache_size", float64(cacheSize))
 
 	totalDataSizeSent := 0
 	tracesSent := generics.NewSet[string]()
@@ -483,7 +483,7 @@ func (i *InMemCollector) redistributeTraces(ctx context.Context) {
 	redistrubutionStartTime := i.Clock.Now()
 
 	defer func() {
-		i.Metrics.Histogram("collector_redistribute_traces_duration_ms", i.Clock.Now().Sub(redistrubutionStartTime).Milliseconds())
+		i.Metrics.Histogram("collector_redistribute_traces_duration_ms", float64(i.Clock.Now().Sub(redistrubutionStartTime).Milliseconds()))
 		span.End()
 	}()
 
@@ -571,7 +571,7 @@ func (i *InMemCollector) redistributeTraces(ctx context.Context) {
 		"hostname":              i.hostname,
 	})
 
-	i.Metrics.Gauge("trace_forwarded_on_peer_change", len(forwardedTraces)+len(emptyTraces))
+	i.Metrics.Gauge("trace_forwarded_on_peer_change", float64(len(forwardedTraces)+len(emptyTraces)))
 
 	// remove all redistributed traces from the cache if we are in traces locality concentrated mode
 	if i.Config.GetCollectionConfig().TraceLocalityEnabled() {
@@ -589,7 +589,7 @@ func (i *InMemCollector) sendExpiredTracesInCache(ctx context.Context, now time.
 	ctx, span := otelutil.StartSpan(ctx, i.Tracer, "sendExpiredTracesInCache")
 	startTime := time.Now()
 	defer func() {
-		i.Metrics.Histogram("collector_send_expired_traces_in_cache_dur_ms", time.Since(startTime).Milliseconds())
+		i.Metrics.Histogram("collector_send_expired_traces_in_cache_dur_ms", float64(time.Since(startTime).Milliseconds()))
 		span.End()
 	}()
 
@@ -622,8 +622,8 @@ func (i *InMemCollector) sendExpiredTracesInCache(ctx context.Context, now time.
 	})
 
 	dur := time.Now().Sub(startTime)
-	i.Metrics.Gauge("collector_expired_traces_missing_decisions", len(expiredTraces))
-	i.Metrics.Gauge("collector_expired_traces_orphans", orphanTraceCount)
+	i.Metrics.Gauge("collector_expired_traces_missing_decisions", float64(len(expiredTraces)))
+	i.Metrics.Gauge("collector_expired_traces_orphans", float64(orphanTraceCount))
 
 	span.SetAttributes(attribute.Int("num_traces_to_expire", len(traces)), attribute.Int64("take_expired_traces_duration_ms", dur.Milliseconds()))
 
@@ -1473,7 +1473,7 @@ func (i *InMemCollector) processTraceDecisions(msg string, decisionType decision
 		return
 	}
 
-	i.Metrics.Count(fmt.Sprintf("%s_decisions_received", decisionType.String()), len(decisions))
+	i.Metrics.Count(fmt.Sprintf("%s_decisions_received", decisionType.String()), int64(len(decisions)))
 
 	if len(decisions) == 0 {
 		return
@@ -1597,7 +1597,7 @@ func (i *InMemCollector) IsMyTrace(traceID string) (sharder.Shard, bool) {
 func (i *InMemCollector) publishTraceDecision(ctx context.Context, td TraceDecision) {
 	start := time.Now()
 	defer func() {
-		i.Metrics.Histogram("collector_publish_trace_decision_dur_ms", time.Since(start).Milliseconds())
+		i.Metrics.Histogram("collector_publish_trace_decision_dur_ms", float64(time.Since(start).Milliseconds()))
 	}()
 
 	_, span := otelutil.StartSpanWith(ctx, i.Tracer, "publishTraceDecision", "decision", td.Kept)
@@ -1696,7 +1696,7 @@ func (i *InMemCollector) sendDecisions(decisionChan <-chan TraceDecision, interv
 
 		// Send the batch if ready
 		if send && len(decisions) > 0 {
-			i.Metrics.Histogram(metricName, len(decisions))
+			i.Metrics.Histogram(metricName, float64(len(decisions)))
 
 			// Copy current batch to process
 			decisionsToProcess := make([]TraceDecision, len(decisions))
